Coffee_Shop_Chatbot/python_code/api/requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
openai==1.55.3
runpod==1.7.1
pinecone==5.3.1
httpx==0.27.2
typing-extensions>=4.11.0,<5.0.0